+++ pkg/db/model.go
@@ -220,8 +220,8 @@ type BrokerProduct struct {
 	ConnectionCode   string  `json:"connectionCode"`
 	Symbol           string  `json:"symbol"`
 	Name             string  `json:"name"`
-	PointValue       float32 `json:"pointValue"`
-	CostPerOperation float32 `json:"costPerOperation"`
+	PointValue       float64 `json:"pointValue"`
+	CostPerOperation float64 `json:"costPerOperation"`
 	CurrencyCode     string  `json:"currencyCode"`
 }
